Fix more integer overflow issues

See merge request octopus-code/octopus!1273

+++ src/math/checksum_interface.F90
@@ -25,7 +25,7 @@ module checksum_interface_oct_m
     subroutine checksum_calculate(algorithm, narray, array, checksum)
       implicit none
       integer,    intent(in)  :: algorithm
-      integer,    intent(in)  :: narray
+      integer(8), intent(in)  :: narray
       integer,    intent(in)  :: array
       integer(8), intent(out) :: checksum
     end subroutine checksum_calculate
